feat(model-runner): migrate to v0.6.0 manifest + decorator API; split into modular files
Restructures the model-runner app from the v0.5 two-file monolith
(entry_deployment.py 2256 lines, runtime_deployment.py 377 lines) into
five smaller, single-responsibility files plus the manifest:

  apps/model-runner/
  ├── manifest.yaml             # v0.6.0, entry: entry:EntryApp
  ├── README.md
  ├── entry.py                  # @bioengine.app EntryApp (RPC surface)
  ├── runtime.py                # @bioengine.app RuntimeApp (GPU executor)
  └── model_cache/
      ├── __init__.py
      ├── package.py            # BioimageioPackage (per-use lock)
      └── cache.py              # ModelCache (LRU + downloader)

The v0.5 → v0.6 transforms are the same shape used in the demo-app and
composition-demo migrations:

* ``@serve.deployment(ray_actor_options={...})`` → ``@bioengine.app(...)``
  with flat ``num_cpus`` / ``num_gpus`` / ``memory_mb`` / ``pip`` args;
  autoscaling / health-check / graceful-shutdown config passes through
  to the underlying ``serve.deployment`` via ``**serve_deployment_kwargs``.
* ``@schema_method`` → ``@bioengine.method`` (with a small app-local
  ``_arbitrary_types_method`` helper for ``infer``, which has ``np.ndarray``
  in its signature).
* ``async def async_init`` / ``check_health`` → wrapped in
  ``@bioengine.async_init`` / ``@bioengine.health_check`` with private
  underscore names.
* ``runtime_deployment: DeploymentHandle`` → ``runtime: RuntimeApp``
  (the v0.6 composition graph is declared by type hints).
* Every ``self.runtime_deployment.X.remote(...)`` becomes
  ``await self.runtime.X(...)`` — the ``BioEngineRuntimeHandle`` proxy
  hides ``.remote()``.
* ``self.runtime.check_health()`` (a cross-deployment liveness ping
  in the v0.5 design) becomes ``self.runtime.ping()`` — ``check_health``
  is now framework-owned and not user-callable.
* ``serve.get_replica_context()`` is moved to a lazy ``from ray import
  serve`` inside ``__init__`` (the module no longer imports
  ``ray.serve`` at top level).

``RuntimeApp`` keeps the full multiplexed-pipeline + memory-accounting
logic, just with the v0.6 decorators. The pinned dep set (REQUIREMENTS)
is identical to the v0.5 runtime_deployment.py.

``ModelCache`` and ``BioimageioPackage`` are unchanged in behaviour;
they only moved files. ``EntryApp`` imports them from ``model_cache``.

Manifest bumps version 1.0.7 → 1.1.0.
