+++ apps/model-runner/manifest.yaml
@@ -1,20 +1,17 @@
-authors:
-  - { affiliation: KTH, github_user: nilsmechtel, name: Nils Mechtel }
-covers: []
-description: Runs and compares BioImage.IO models through the BioEngine service. It supports microscopy image inference, model search, and validation.
-documentation: "README.md"
-tutorial: "https://imjoy-notebook.netlify.app/lab/index.html?load=https://raw.githubusercontent.com/aicell-lab/bioengine/refs/heads/main/apps/model-runner/tutorial_model_runner.ipynb&open=1"
-format_version: 0.5.0
-git_repo: ""
+name: Model Runner
 id: model-runner
 id_emoji: "🦒"
-license: MIT
-links: []
+description: Runs and compares BioImage.IO models through the BioEngine service. It supports microscopy image inference, model search, and validation.
+documentation: README.md
+tutorial: "https://imjoy-notebook.netlify.app/lab/index.html?load=https://raw.githubusercontent.com/aicell-lab/bioengine/refs/heads/main/apps/model-runner/tutorial_model_runner.ipynb&open=1"
+authors:
+  - {affiliation: KTH, github_user: nilsmechtel, name: Nils Mechtel}
 maintainers: []
-name: Model Runner
+license: MIT
+format_version: 0.6.0
+git_repo: https://github.com/aicell-lab/bioengine
+links: []
 tags: []
 type: ray-serve
-version: 1.0.7
-deployments:
-  - entry_deployment:EntryDeployment
-  - runtime_deployment:RuntimeDeployment
+version: 1.1.0
+entry: entry:EntryApp
