+++ apps/model-runner/model_cache/__init__.py
@@ -1,0 +1,10 @@
+"""LRU model cache for bioimage.io model packages.
+
+See :mod:`model_cache.cache` for the cache implementation and
+:mod:`model_cache.package` for the per-use lock context manager.
+"""
+
+from model_cache.cache import ModelCache
+from model_cache.package import BioimageioPackage
+
+__all__ = ["BioimageioPackage", "ModelCache"]
